requirements.txt
pandas
scikit-learn
bentoml
pyjwt
pydantic
starlette
numpy
requests
